Enhance README with team details and mission statement
🏆 Team 3 - Round 4 | Digital Egypt Pioneers Initiative (DEDI)

Track: AI & Data Science - Microsoft Data Engineer Specialization

Our Mission: We are on a journey to master the use of Microsoft tools for data management and processing. Our goal is to build advanced skills in SQL, the Microsoft Azure platform, and DevOps roles to design robust, secure, and scalable data solutions.

Our Vision: To be a distinguished team that collaborates to maximize the benefit of the initiative, supporting each other to achieve our professional goals as top-tier Data Engineers.

Together.. We Build the Future! 🔥